feat: drop support for laminas-eventmanager v2 and laminas-servicemanager v2
