+++ composer.json
@@ -26,7 +26,7 @@
     },
     "require": {
         "php": "^7.3 || ~8.0.0",
-        "laminas/laminas-eventmanager": "^2.6.2 || ^3.0",
+        "laminas/laminas-eventmanager": "^3.0",
         "laminas/laminas-stdlib": "^3.2.1",
         "laminas/laminas-zendframework-bridge": "^1.0"
     },
@@ -36,7 +36,7 @@
         "laminas/laminas-coding-standard": "~1.0.0",
         "laminas/laminas-db": "^2.7",
         "laminas/laminas-http": "^2.5.4",
-        "laminas/laminas-servicemanager": "^2.7.5 || ^3.0.3",
+        "laminas/laminas-servicemanager": "^3.0.3",
         "laminas/laminas-validator": "^2.6",
         "mongodb/mongodb": "^1.0.1",
         "php-mock/php-mock-phpunit": "^1.1.2 || ^2.0",
